@@ -1,12 +1,12 @@
 package com.mycompany.myapp;
 
+import jakarta.annotation.Generated;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import javax.annotation.Generated;
 
-@Generated(value = "JHipster", comments = "Generated by JHipster 7.9.4")
+@Generated(value = "JHipster", comments = "Generated by JHipster 8.7.3")
 @Retention(RetentionPolicy.SOURCE)
 @Target({ ElementType.TYPE })
 public @interface GeneratedByJHipster {
