@@ -4,6 +4,7 @@ import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.pattern.CompositeConverter;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import org.slf4j.Marker;
 import org.slf4j.MarkerFactory;
@@ -12,11 +13,22 @@ import org.springframework.boot.ansi.AnsiElement;
 import org.springframework.boot.ansi.AnsiOutput;
 import org.springframework.boot.ansi.AnsiStyle;
 
+/**
+ * Log filter to prevent attackers from forging log entries by submitting input containing CRLF characters.
+ * CRLF characters are replaced with a red colored _ character.
+ *
+ * @see <a href="https://owasp.org/www-community/attacks/Log_Injection">Log Forging Description</a>
+ * @see <a href="https://github.com/jhipster/generator-jhipster/issues/14949">JHipster issue</a>
+ */
 public class CRLFLogConverter extends CompositeConverter<ILoggingEvent> {
 
     public static final Marker CRLF_SAFE_MARKER = MarkerFactory.getMarker("CRLF_SAFE");
 
-    private static final String[] SAFE_LOGGERS = { "org.hibernate" };
+    private static final String[] SAFE_LOGS = {
+        "org.hibernate",
+        "org.springframework.boot.autoconfigure",
+        "org.springframework.boot.diagnostics",
+    };
     private static final Map<String, AnsiElement> ELEMENTS;
 
     static {
@@ -34,7 +46,8 @@ public class CRLFLogConverter extends CompositeConverter<ILoggingEvent> {
     @Override
     protected String transform(ILoggingEvent event, String in) {
         AnsiElement element = ELEMENTS.get(getFirstOption());
-        if ((event.getMarker() != null && event.getMarker().contains(CRLF_SAFE_MARKER)) || isLoggerSafe(event)) {
+        List<Marker> markers = event.getMarkerList();
+        if ((markers != null && !markers.isEmpty() && markers.get(0).contains(CRLF_SAFE_MARKER)) || isLoggerSafe(event)) {
             return in;
         }
         String replacement = element == null ? "_" : toAnsiString("_", element);
@@ -42,7 +55,7 @@ public class CRLFLogConverter extends CompositeConverter<ILoggingEvent> {
     }
 
     protected boolean isLoggerSafe(ILoggingEvent event) {
-        for (String safeLogger : SAFE_LOGGERS) {
+        for (String safeLogger : SAFE_LOGS) {
             if (event.getLoggerName().startsWith(safeLogger)) {
                 return true;
             }
