@@ -1,13 +1,13 @@
 package com.mycompany.myapp.web.rest.errors;
 
 import java.net.URI;
-import java.util.HashMap;
-import java.util.Map;
-import org.zalando.problem.AbstractThrowableProblem;
-import org.zalando.problem.Status;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.ErrorResponseException;
+import tech.jhipster.web.rest.errors.ProblemDetailWithCause;
+import tech.jhipster.web.rest.errors.ProblemDetailWithCause.ProblemDetailWithCauseBuilder;
 
 @SuppressWarnings("java:S110") // Inheritance tree of classes should not be too deep
-public class BadRequestAlertException extends AbstractThrowableProblem {
+public class BadRequestAlertException extends ErrorResponseException {
 
     private static final long serialVersionUID = 1L;
 
@@ -20,7 +20,17 @@ public class BadRequestAlertException extends AbstractThrowableProblem {
     }
 
     public BadRequestAlertException(URI type, String defaultMessage, String entityName, String errorKey) {
-        super(type, defaultMessage, Status.BAD_REQUEST, null, null, null, getAlertParameters(entityName, errorKey));
+        super(
+            HttpStatus.BAD_REQUEST,
+            ProblemDetailWithCauseBuilder.instance()
+                .withStatus(HttpStatus.BAD_REQUEST.value())
+                .withType(type)
+                .withTitle(defaultMessage)
+                .withProperty("message", "error." + errorKey)
+                .withProperty("params", entityName)
+                .build(),
+            null
+        );
         this.entityName = entityName;
         this.errorKey = errorKey;
     }
@@ -33,10 +43,7 @@ public class BadRequestAlertException extends AbstractThrowableProblem {
         return errorKey;
     }
 
-    private static Map<String, Object> getAlertParameters(String entityName, String errorKey) {
-        Map<String, Object> parameters = new HashMap<>();
-        parameters.put("message", "error." + errorKey);
-        parameters.put("params", entityName);
-        return parameters;
+    public ProblemDetailWithCause getProblemDetailWithCause() {
+        return (ProblemDetailWithCause) this.getBody();
     }
 }
