@@ -9,9 +9,7 @@ import org.testcontainers.containers.output.Slf4jLogConsumer;
 
 public class MysqlTestContainer implements SqlTestContainer {
 
-    private static final Logger log = LoggerFactory.getLogger(MysqlTestContainer.class);
-    private long memoryInBytes = 100 * 1024 * 1024;
-    private long memorySwapInBytes = 200 * 1024 * 1024;
+    private static final Logger LOG = LoggerFactory.getLogger(MysqlTestContainer.class);
 
     private MySQLContainer<?> mysqlContainer;
 
@@ -25,15 +23,11 @@ public class MysqlTestContainer implements SqlTestContainer {
     @Override
     public void afterPropertiesSet() {
         if (null == mysqlContainer) {
-            mysqlContainer =
-                new MySQLContainer<>("mysql:8.0.30-debian")
-                    .withDatabaseName("jhipster")
-                    .withTmpFs(Collections.singletonMap("/testtmpfs", "rw"))
-                    .withLogConsumer(new Slf4jLogConsumer(log))
-                    .withReuse(true)
-                    .withPrivilegedMode(true)
-                    .withConfigurationOverride("testcontainers/mysql")
-                    .withCreateContainerCmdModifier(cmd -> cmd.getHostConfig().withMemory(memoryInBytes).withMemorySwap(memorySwapInBytes));
+            mysqlContainer = new MySQLContainer<>("mysql:9.1.0")
+                .withDatabaseName("jhipster")
+                .withTmpFs(Collections.singletonMap("/testtmpfs", "rw"))
+                .withLogConsumer(new Slf4jLogConsumer(LOG))
+                .withReuse(true);
         }
         if (!mysqlContainer.isRunning()) {
             mysqlContainer.start();
