@@ -7,8 +7,12 @@ import org.springframework.web.bind.annotation.RestController;
 public class WebConfigurerTestController {
 
     @GetMapping("/api/test-cors")
-    public void testCorsOnApiPath() {}
+    public void testCorsOnApiPath() {
+        // empty method
+    }
 
     @GetMapping("/test/test-cors")
-    public void testCorsOnOtherPath() {}
+    public void testCorsOnOtherPath() {
+        // empty method
+    }
 }
